fix: Merge Output Executor with audio
